Converted balance to floating point

+++ src/models.rs
@@ -13,19 +13,19 @@ pub struct Account {
     pub id: i32,
     pub acct_name: String,
     pub acct_type: String,
-    pub debit_credit: i32,
+    pub debit_credit: f32,
 }
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct TransactionSplit {
-    pub amount: i32,
+    pub amount: f32,
     pub account: String,
 }
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub struct Transaction {
     pub date: String,
-    pub debit_credit: i32,
+    pub debit_credit: f32,
     pub acct_name: String,
     pub acct_type: String,
     pub acct_offset_name: String,
